Backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-recurrence==1.11.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
setuptools==76.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
wheel==0.45.1
whitenoise==6.9.0
ratelimit==2.2.1
